+++ src/table/index.js
@@ -2,3 +2,4 @@
 export { PagedTable } from "./paged-table";
 export { SimpleTable } from "./simple-table";
 export { LargeSimpleTable } from "./large-simple-table";
+export { ExportTableButton } from "./components/export-button";